feat(oracle): add test suite and snapshots for issues #33 #34 #35 #36
Issue #34 - [Test/Oracle] Write unit tests for init and double-init guard:
- Create contracts/oracle/src/test.rs with #![cfg(test)]
- Add test_init: registers contract, calls init, asserts success
- Add test_init_already_initialized: calls init twice, asserts Error::AlreadyInitialized
- Wire mod test; at the bottom of lib.rs, removing inline mod tests block

Issue #35 - [Test/Oracle] Write unit tests for submit_request uniqueness and TTL:
- Add test_submit_request_generates_unique_ids: submits two requests, asserts IDs are 1 and 2
- Add test_submit_request_stores_pending_in_temporary_storage: asserts state is Pending and TTL > 0

Issue #36 - [Test/Oracle] Write unit tests for verify_attestation:
- Add test_verify_attestation_success: valid Ed25519 signature, asserts Ok(())
- Add test_verify_attestation_invalid_signature: bad signature, asserts error
- Add test_verify_attestation_unauthorized_signer: unregistered provider, asserts Error::UnauthorizedSigner
- Add test_verify_attestation_not_initialized: call without init, asserts Error::RegistryNotConfigured

Issue #33 - [Contract/All] Add test snapshot files for all contract tests:
- soroban-sdk testutils feature already enabled in all three Cargo.toml files
- Add 8 snapshot JSON files under contracts/oracle/test_snapshots/test/
- Add 6 snapshot JSON files under contracts/provenance/test_snapshots/tests/
- Add 6 snapshot JSON files under contracts/registry/test_snapshots/tests/
